refactor(trajectory): decompose monolith into modular trajectory engine
