+++ src/movement_optimizer/trajectory/__init__.py
@@ -1,0 +1,16 @@
+"""Trajectory module extracted from monolith."""
+from __future__ import annotations
+
+from .cache import SolutionCache as SolutionCache
+from .optimizer import TrajectoryOptimizer as TrajectoryOptimizer
+from .result import CancelledError as CancelledError
+from .result import OptimizationResult as OptimizationResult
+from .result import ProgressReport as ProgressReport
+
+__all__ = [
+    "CancelledError",
+    "OptimizationResult",
+    "ProgressReport",
+    "SolutionCache",
+    "TrajectoryOptimizer",
+]
